image_retreiver_from_captions/requirements.txt
torch
torchvision
pillow
sentence-transformers
transformers
faiss-cpu
tqdm
numpy
qdrant-client
opencv-python
